About 30% true code coverage reached
